Refactor components to design system and clean styles
