+++ src/components/Hero.tsx
@@ -1,50 +1,72 @@
-
-import { GlassButton } from "@/components/ui/glass-button";
+import { Section, Container, Typography, Button, BackgroundImage } from '@/components/design-system';
 
 export default function Hero() {
   return (
-    <section className="relative h-screen flex items-center justify-center overflow-hidden">
-      <div 
-        className="absolute inset-0 bg-cover bg-center bg-no-repeat"
-        style={{
-          backgroundImage: `url('https://images.unsplash.com/photo-1518005020951-eccb494ad742?q=80&w=2564&auto=format&fit=crop')`,
-        }}
+    <Section 
+      fullHeight 
+      padding="none"
+      as="main"
+      className="flex items-center justify-center"
+    >
+      <BackgroundImage
+        src="https://images.unsplash.com/photo-1518005020951-eccb494ad742?q=80&w=2564&auto=format&fit=crop"
+        alt="Modern luxury apartment interior in Fulton Market"
+        overlayOpacity={70}
+        grain
+        className="absolute inset-0"
       />
       
-      <div className="absolute inset-0 cinematic-gradient" />
-      <div className="absolute inset-0 bg-grain opacity-20" />
-      
-      <div className="relative z-10 text-center px-4 max-w-4xl mx-auto">
-        <h1 className="font-serif text-5xl md:text-7xl lg:text-8xl font-light text-white mb-6 hero-text tracking-tight">
-          The apartment experience{" "}
-          <span className="italic gold-accent">that finds you.</span>
-        </h1>
+      <Container size="lg" className="text-center">
+        <Typography
+          variant="h1"
+          family="serif"
+          weight="light"
+          align="center"
+          balance
+          className="mb-6 text-shadow-luxury"
+        >
+          The apartment marketplace{" "}
+          <span className="italic text-amber-200/90">
+            that finds you.
+          </span>
+        </Typography>
         
-        <p className="text-xl md:text-2xl text-neutral-300 mb-12 font-light tracking-wide max-w-2xl mx-auto">
-          A premium rental network connecting you to Fulton Market's most sought-after addresses and hidden gems.
-        </p>
+        <Typography
+          variant="body-lg"
+          color="secondary"
+          weight="light"
+          align="center"
+          className="mb-12 max-w-2xl mx-auto"
+        >
+          A premium rental network connecting you to Fulton Market's most 
+          sought-after addresses and hidden gems.
+        </Typography>
         
         <div className="flex flex-col sm:flex-row gap-6 justify-center items-center">
-          <GlassButton 
+          <Button 
+            variant="primary"
             size="lg"
-            className="glass-primary"
+            className="min-w-48"
           >
             Get Early Access
-          </GlassButton>
+          </Button>
           
-          <GlassButton 
+          <Button 
+            variant="glass"
             size="lg"
+            className="min-w-48"
           >
             Learn More
-          </GlassButton>
+          </Button>
         </div>
-      </div>
+      </Container>
       
+      {/* Scroll indicator */}
       <div className="absolute bottom-8 left-1/2 transform -translate-x-1/2 animate-bounce">
         <div className="w-6 h-10 border-2 border-white/30 rounded-full flex justify-center">
           <div className="w-1 h-3 bg-white/50 rounded-full mt-2 animate-pulse" />
         </div>
       </div>
-    </section>
+    </Section>
   );
 }
