+++ src/components/Invitation.tsx
@@ -1,32 +1,47 @@
+import { Section, Container, Typography, BackgroundImage } from '@/components/design-system';
 
 export default function Invitation() {
   return (
-    <section className="relative min-h-screen flex items-center">
-      <div 
-        className="absolute inset-0 bg-cover bg-center bg-no-repeat"
-        style={{
-          backgroundImage: `url('https://images.unsplash.com/photo-1618221195710-dd6b41faaea6?q=80&w=2400&auto=format&fit=crop')`,
-        }}
+    <Section 
+      fullHeight 
+      padding="xl"
+      className="flex items-center"
+    >
+      <BackgroundImage
+        src="https://images.unsplash.com/photo-1618221195710-dd6b41faaea6?q=80&w=2400&auto=format&fit=crop"
+        alt="Elegant modern living space with premium finishes"
+        overlayOpacity={70}
+        className="absolute inset-0"
       />
       
-      <div className="absolute inset-0 bg-black/70" />
-      
-      <div className="relative z-10 container py-20">
-        <div className="max-w-4xl mx-auto text-center">
-          <div className="space-y-12">
-            <h2 className="font-serif text-4xl md:text-6xl lg:text-7xl font-light text-white tracking-tight leading-tight">
-              No listings, no noise.
-            </h2>
-            
-            <div className="w-32 h-px bg-amber-200/60 mx-auto" />
-            
-            <p className="text-2xl md:text-3xl text-neutral-300 font-light max-w-3xl mx-auto leading-relaxed">
-              A curated rental experience built for those who value{" "}
-              <span className="gold-accent italic">quality over quantity.</span>
-            </p>
-          </div>
+      <Container size="lg">
+        <div className="max-w-4xl mx-auto text-center space-y-12">
+          <Typography
+            variant="h2"
+            family="serif"
+            weight="light"
+            align="center"
+            balance
+            className="text-shadow-luxury"
+          >
+            No listings, no noise.
+          </Typography>
+          
+          {/* Decorative line */}
+          <div className="w-32 h-px bg-amber-200/60 mx-auto" />
+          
+          <Typography
+            variant="body-lg"
+            color="secondary"
+            weight="light"
+            align="center"
+            className="max-w-3xl mx-auto"
+          >
+            A curated rental experience built for those who value{" "}
+            <span className="italic text-amber-200/90">quality over quantity.</span>
+          </Typography>
         </div>
-      </div>
-    </section>
+      </Container>
+    </Section>
   );
 }
