+++ src/components/design-system/index.ts
@@ -3,7 +3,7 @@ export { Section } from './Section';
 export { Container } from './Container';
 export { Typography } from './Typography';
 export { Button } from './Button';
-export { Input } from './Input';
+export { Input, Textarea } from './Input';
 export { Accordion } from './Accordion';
 export { BackgroundImage } from './BackgroundImage';
 export { FeatureGrid } from './FeatureGrid';
@@ -14,7 +14,7 @@ export type { SectionProps } from './Section';
 export type { ContainerProps } from './Container';
 export type { TypographyProps } from './Typography';
 export type { ButtonProps } from './Button';
-export type { InputProps } from './Input';
+export type { InputProps, TextareaProps } from './Input';
 export type { AccordionProps } from './Accordion';
 export type { BackgroundImageProps } from './BackgroundImage';
 export type { FeatureGridProps } from './FeatureGrid';
